Music packet adjustment. Could have sworn client was using larger than 8 bits for that, but I was apparently seeing the blockID and SongID together and mistook it for the songID alone.

+++ src/map/packets/change_music.cpp
@@ -25,14 +25,15 @@
 
 #include "change_music.h"
 
-CChangeMusicPacket::CChangeMusicPacket(uint8 BlockID, uint32 MusicTrackID)
+CChangeMusicPacket::CChangeMusicPacket(uint8 BlockID, uint8 MusicTrackID)
 {
     // Block IDs:
-    // 0 Background Music
-    // 1 Alternate Background Music (SummerFest etc?)
+    // 0 Background Music (Day time, 7:00 -> 18:00)
+    // 1 Background Music (Night time, 18:00 -> 7:00)
     // 2 SoloBattle Music
     // 3 Party Battle Music
     // 4 Chocobo Music
+    // More types likely exist
 
     this->type = 0x5F;
     this->size = 0x04;
